src/api/v1/endpoints/auth.py
"""
Authentication endpoints
"""

from fastapi import APIRouter, HTTPException, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel
from typing import Dict, Any
import logging

logger = logging.getLogger(__name__)

router = APIRouter()
security = HTTPBearer()


class TokenRequest(BaseModel):
    """Token request model"""
    username: str
    password: str


class TokenResponse(BaseModel):
    """Token response model"""
    access_token: str
    token_type: str = "bearer"
    expires_in: int = 3600


@router.post("/token", response_model=TokenResponse)
async def create_access_token(request: TokenRequest) -> TokenResponse:
    """Create access token for API authentication"""
    # TODO: Implement actual authentication logic
    # For now, return a dummy token
    logger.info(f"Token request for user: {request.username}")
    
    if request.username == "admin" and request.password == "admin":
        return TokenResponse(
            access_token="dummy_token_12345",
            token_type="bearer",
            expires_in=3600
        )
    
    raise HTTPException(status_code=401, detail="Invalid credentials")


@router.get("/verify")
async def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)) -> Dict[str, Any]:
    """Verify API token"""
    # TODO: Implement actual token verification
    logger.info(f"Token verification request: {credentials.credentials[:10]}...")
    
    return {
        "valid": True,
        "user_id": "user_123",
        "scopes": ["read", "write"]
    }

