requirements-simple.txt
# Core dependencies (Python 3.13 compatible)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
litellm>=1.44.22
pydantic-settings>=2.1.0

# Database (using asyncpg instead of psycopg2 for better Python 3.13 support)
sqlalchemy>=2.0.23
alembic>=1.13.1
asyncpg>=0.29.0

# Caching and Redis
redis>=5.0.1

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP and API
httpx>=0.25.2
requests>=2.31.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Environment and Configuration
python-dotenv>=1.0.0

# Testing (dev dependencies)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
